// src/system-library/types.ts
import { SymbolInfo } from "../symbol-indexer";

export type SystemContainer =
  | 'Data7'
  | 'System'
  | 'System.Classes'
  | 'System.Classes.TObject'
  | 'System.Classes.TPersistent'
  | 'Environment'
  | 'IO'
  | 'Collections'
  | 'XML'
  | 'Forms'
  | 'SQL'
  | 'Drawing'
  | 'Excel'
  | 'IOUtils'
  | 'Directory'
  | 'File'
  | 'ZipFile'
  | 'StringList'
  | 'TXMLDocument'
  | 'IXMLNode'
  | 'MessageBox'
  | 'Form'
  | 'Grid'
  | 'Imagem'
  | 'TMargins'
  | 'TAlign'
  | 'TWinControl'
  | 'CustomControl'
  | 'StaticText'
  | 'TBoundLabel'
  | 'TLabeledEdit'
  | 'TAlignment'
  | 'Connection'
  | 'Command'
  | 'TField'
  | 'TFDParam'
  | 'TCanvas'
  | 'TPen'
  | 'String'
  | 'Integer'
  | 'Double'
  | 'Boolean'
  | 'TDateTime'
  | 'Date'
  | 'Single'
  | 'TObject'
  | 'TPersistent'
  | 'TStrings'
  | 'TStringList'
  | 'THTTP'
  | 'TJSONObject'
  | 'TJSONArray'
  | 'Forms.Form'
  | 'Forms.Grid'
  | 'Forms.Imagem';

export interface SystemSymbolInfo extends Omit<SymbolInfo, 'containerName'> {
  containerName?: SystemContainer;
}